Extract shared team request helpers
